refactor: extract registering middleware into own function
