+++ src/server/server.ts
@@ -1,18 +1,8 @@
-import { json } from 'body-parser'
-import session from 'express-session'
-import { expressMiddleware } from '@apollo/server/express4'
-import cors from 'cors'
-import { createExpressApp } from './express/express'
+import { createExpressApp, registerMiddleware } from './express/express'
 import { createApolloServer } from './graphql/graphql'
 import { createHttpServer } from './http/http'
 import { createWsServer } from './ws/ws'
 import config from '../config'
-
-const corsOptions = {
-	origin: config.CLIENT_URL,
-	credentials: true,
-	optionsSuccessStatus: 204,
-}
 
 export async function startChessServer() {
 	const app = createExpressApp()
@@ -22,23 +12,7 @@ export async function startChessServer() {
 
 	await apolloServer.start()
 
-	app.use(cors(corsOptions))
-	app.use(
-		session({
-			secret: config.SESSION_SECRET,
-			resave: false,
-			saveUninitialized: true,
-		}),
-	)
-	app.use(
-		'/graphql',
-		json(),
-		expressMiddleware(apolloServer, {
-			context: async (data: any) => {
-				return { sessionId: data.req.sessionID }
-			},
-		}),
-	)
+	registerMiddleware({ expressApp: app, apolloServer })
 
 	return new Promise((resolve) =>
 		httpServer.listen(config.PORT, () => resolve(true)),
